requirements.txt
psycopg[binary]
python-dotenv
fastapi[standard]
structlog
pytest
fastapi 
uvicorn
approvaltests
pytest-approvaltests
coverage
ruff 
motor
alembic
asyncio
httpx
pytest-asyncio
pytest-cov
sqlalchemy

